Exception in case of id not found at the http verbs

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.alura.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -62,26 +63,42 @@ public class TopicosController {
 	}
 
 	@GetMapping("/{id}")
-	public DetalhesDoTopicoDTO detalhar(@PathVariable Long id) {
-		Topico topico = topicoRepository.findById(id).get();
-		return new DetalhesDoTopicoDTO(topico);
+	public ResponseEntity<DetalhesDoTopicoDTO> detalhar(@PathVariable Long id) {
+		Optional<Topico> topico = topicoRepository.findById(id);
+		if (topico.isPresent()) {
+			return ResponseEntity.ok(new DetalhesDoTopicoDTO(topico.get()));
+		}
+
+		return ResponseEntity.notFound().build();
 	}
 
 	@PutMapping("/{id}")
-	@Transactional // O Spring data diz que todo método que tiver operação de escrita, deve colocar o @Transactional
+	@Transactional // O Spring data diz que todo método que tiver operação de escrita, deve colocar
+					// o @Transactional
 	public ResponseEntity<TopicoDTO> atualizar(@PathVariable Long id,
 			@RequestBody @Valid AtualizacaoTopicoForm topicoForm) {
-		Topico topico = topicoForm.atualizar(id, topicoRepository);
 
-		return ResponseEntity.ok(new TopicoDTO(topico));
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if (optional.isPresent()) {
+			Topico topico = topicoForm.atualizar(id, topicoRepository);
+			return ResponseEntity.ok(new TopicoDTO(topico));
+		}
+
+		return ResponseEntity.notFound().build();
 
 	}
 
 	@DeleteMapping("/{id}")
 	@Transactional
 	public ResponseEntity<?> remover(@PathVariable Long id) {
-		topicoRepository.deleteById(id);
-		return ResponseEntity.ok().build();
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if (optional.isPresent()) {
+			topicoRepository.deleteById(id);
+			return ResponseEntity.ok().build();
+
+		}
+
+		return ResponseEntity.notFound().build();
 	}
 
 }
